+++ src/core/component/engines/engine.ts
@@ -7,16 +7,3 @@
  */
 
 export * from 'core/component/engines/vue3';
-export * from 'core/component/engines/interface';
-
-export { VNode } from 'core/component/engines/interface';
-
-export {
-
-	CreateAppFunction,
-
-	Directive,
-	DirectiveArguments,
-	ObjectDirective
-
-} from 'core/component/engines/vue3';