+++ src/core/component/engines/vue3/index.ts
@@ -16,7 +16,6 @@ import 'core/component/engines/vue3/config';
 
 export * from 'core/component/engines/vue3/const';
 export * from 'core/component/engines/vue3/component';
-export * from 'core/component/engines/vue3/interface';
 
 export { Vue as ComponentEngine };
 export { Vue as default };